+++ resolvers/webpack/CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](http://keepachangelog.com).
 
 ## Unreleased
+### Added
+- support for config-as-function ([#533], thanks [@grahamb])
 
 ## 0.5.1 - 2016-08-11
 ### Fixed
@@ -58,6 +60,7 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 - `interpret` configs (such as `.babel.js`).
   Thanks to [@gausie] for the initial PR ([#164], ages ago! 😅) and [@jquense] for tests ([#278]).
 
+[#533]: https://github.com/benmosher/eslint-plugin-import/pull/533
 [#413]: https://github.com/benmosher/eslint-plugin-import/pull/413
 [#377]: https://github.com/benmosher/eslint-plugin-import/pull/377
 [#363]: https://github.com/benmosher/eslint-plugin-import/pull/363
@@ -82,3 +85,4 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 [@Satyam]: https://github.com/Satyam
 [@Rogeres]: https://github.com/Rogeres
 [@Kovensky]: https://github.com/Kovensky
+[@grahamb]: https://github.com/grahamb